Rewrote and simplified service ticket validation scenarios.
Specifically, the scenarios were edited to revolve around the idea of a
CAS server, a CAS-protected application, and a user logging into the CAS
server.  The application is (well, will be) written to accept all
requests that contain a valid service ticket and reject all requests
that contain an invalid service ticket.  The service ticket is read from
the query string.

The CAS client will not implement access control, but I expect that its
domainant use will be in such a context, so I edited the scenario to
reflect that.