Added RubyCAS-Server as a development dependency.
The SQLite3 dependency is for ActiveRecord.  The ActiveRecord dependency
is redundant if you just look at the gemspecs; however, for some reason
I haven't yet been able to deduce, a

    bundle install

without the ActiveRecord dependency will, as of the date of this commit,
install ActiveRecord 1.6.0 and ActiveSupport 3.0.3.  This combination
doesn't work with Ruby 1.9.