Stubbed out new CAS server management code.
The new code will adapt the bits we used from the aforementioned
internal project, but will wrap it up with an interface for running and
controlling CAS servers in separate processes.

The complexity of interprocess communication is unfortunately necessary
even at this stage: although there's no reason to prevent Castanet from
working on Ruby 1.8 and 1.9, the current stable release of
RubyCAS-Server does not work on Ruby 1.9, and it doesn't really make
sense to limit Castanet based off of the needs of one of its development
dependencies.