Another big feature restructuring.
This commit reverts the change made in
81bafa1426c9cf3951c705fa6ee126391956811a, and re-introduces the CAS
client-centric language for scenarios.

The reason: it's a real pain to set up Rack applications using
Rack::Builder (which is the easiest way to set up the Castanet client
class in a webapp) in an external process.  Said webapp would have to
run in an external process because it'd be unwieldy to use rack-test to
test an in-process app and something else (say, Mechanize) to talk to
the out-of-process CAS server.