Reimplemented Castanet as a module.  This finishes what was started in the prior commit.
Along the way, also tried to emphasize the one-time-use nature of CAS
service tickets by changing

    service_ticket(ticket).for(service)

to

    ticket = service_ticket(ticket, service)
    ticket.present!

Rationale: The first form does not have any indication that it is not
idempotent.  The second form doesn't explicitly say that service tickets
may only be used once, but I think the verb "present" combined with the
bang communicates that some sort of mutative, non-undoable process is
being done.  The documentation for present! should (hopefully)
communicate the rest.