First cut at Castanet::ProxyTicket.
This implements proxy-ticket retrieval given a PGT and service.  It does
not implement /proxy response parsing or proxy ticket validation.
Furthermore, it does not attempt to deal with PGT validation errors.
The first two are addressed by a Cucumber scenario; the last item needs
a scenario to test it out.