Simplified autoload graph setup for Castanet::Client.
As Castanet is almost always going to be loaded with

    require 'castanet'

we can assume that $LOAD_PATH will include the path to the file
containing the Castanet module, and with that, autoloads for everything
inside of it.  Making that assumption also means that usage of "require
'castanet'" inside Castanet itself will work fine.