Reimplemented SpawnedHttpServer#wait_for using SystemTimer.
This solves one cause of an intermittent problem that deadlocks the
controller and server processes.  The server process' call stack in
gdb on OS X 10.6.6 looks like this at the top when it's locked up:

    #0  0x00007fff88e7fe92 in select$DARWIN_EXTSN ()
    #1  0x000000010002cf51 in rb_thread_select (...) at eval.c:11462
    #2  0x000000010071e401 in ruby_connect (...) at socket.c:1042
    #3  0x00000001007204ba in init_inetsock_internal (...) at socket.c:1264

(Arguments have been elided.  I didn't rigorously check this, but
there's a pretty good chance that this stack trace is the result of
evaluating #http_available?, and the changes seem to bear me out on this
inductive reasoning.  Egg on face if it turns out later investigation
proves me wrong, I guess.)

The rb_thread_select with infinite timeout inside a function named
wait_connectable, which waits for a socket to become ready for writing.
(Note that rb_thread_select itself doesn't block Ruby's green threads.)

Spawning lots of short-lived TCP connections (as the Cucumber tests do)
puts a great deal of localhost -> localhost TCP connections into
TIME_WAIT state.  Under these conditions I have observed under OS X and
Linux that tests that (a) frequently spawn and kill test servers and (b)
do frequent connection testing to see if a port can be bound will be
more prone to what looks like deadlock in initiating connections to
the server.  Unfortunately, I can neither say why that occurs nor
definitively say that it is deadlock, but it certainly acts like it.
More research is required.

In any case, the old retry logic occurred entirely in a single thread.
The ruby_connect call would block the retry and timeout logic (in Ruby)
and therefore never start up nor timeout; and, because the server would
never start up nor timeout, it would never connect to the control pipe
and send a startup notification.  End result: a stalled controller.

This commit solves one of the aspects of the problem by introducing a
timeout mechanism that can be triggered from outside of the currently
executing Ruby thread.  This commit uses SystemTimer, but I think Ruby's
built-in Timeout class should also just work out fine: it doesn't look
like we're hitting the failure cases that SystemTimer is designed to
address. Will check that in a bit.

This commit doesn't eliminate intermittent failures in server startup --
that looks like it has a different cause.  More research required there,
too.  But it's a start.

Future commits will (a) introduce code on the controller side to have a
timeout and/or (b) introduce code on the server side to always close the
pipe to the controller, thus ensuring that the controllers will never
stall waiting for a writer.  Another way to ameliorate the situation
would be to start up all test servers at Cucumber startup, shut them
down at Cucumber termination, and implement purge methods as appropriate
to clear server state, but getting to that point will require some time
to get the purge logic right.