Simplified server subprocess management.
In particular, replaced the HTTP shutdown polling loop with a process
wait, which eliminates the need for HTTP polling and some short-lived
TCP connections, thus reducing the number of connections in TIME_WAIT,
which might increase server startup reliability.

Waiting also means that we no longer need to use Process.detach, because
we will collect the server process' exit status when we exit.