Eliminated Castanet::Error; switched back to `RuntimeError` to indicate a protocol mismatch.  castanet-4.
I'm not sure if there's a reason to differentiate between exception
types for configuration errors.  There might be, in which case this
commit can be reverted.

However, I think it's okay to just treat a configuration error as an
instance of Ruby's default error class.  Because a human operator will
have to step in and fix the problem, the message, not type, is what
matters.