Removed reference to Colin Percival's blog post from `Castanet::Client` documentation.
I originally included the link as a sort of nod to simple design, i.e.
HTTPS totally on or off.  However, that was too simple -- CAS can't work
with HTTPS totally disabled -- so we now have a connection mechanism
that is more complex, but actually works with non-HTTPS connections.

However, inclusion of the article never really made that much sense in
the first place, so I'm going to kill it now.