Update credentials for tests.
The Jasig CAS server ships in a mode that accepts (username, password)
where username = password, and rejects everything else.  (It might also
work with blanks; I haven't tried that.)

Anyway, configuring the Jasig CAS server to do otherwise is terribly
tedious, so I'm going to just go with it.  Servers that are not quite as
massively overengineered can probably be adapted to use this scheme.