Set trust store and its password using Jetty call-things.
HotSpot's argument processing is a real pain to work with.  Splitting
options like this

    "-D", "key=val"

will cause HotSpot to interpret key=val as a class name.  However,
supplying the whole thing as an element in the arglist, i.e.

    "-Dkey=val"

does not set the desired property.

To avoid this mess, this commit sets those properties up in a Jetty
configuration file, which is a pretty ugly hack but is good enough for
this test harness.

    "With all the blood I lost with you
     It drowns the love I thought I knew"