Limit ciphers to be used in Jetty.
The openjdk-{6,7}-jdk packages in Ubuntu seem to have a faulty
implementation of elliptic curve cryptography.
https://bugs.launchpad.net/ubuntu/+source/openjdk-6/+bug/1006776 has
more information.

This commit disables all but one cipher suite that doesn't use ECC.
(For the curious: Diffie-Hellman key exchange, RSA keys in certificate,
AES-128 for symmetric crypto, SHA1 for HMAC.)  The cipher suite was
chosen mostly because it doesn't use ECC, but AFAICT it has no known
major weaknesses.  (Not that that really matters in a test situation.)